requirements.txt
Appium-Python-Client==3.1.0
Appium-Flutter-Finder==0.6.1
requests==2.31.0
python-dotenv==1.0.0